comtroller cocktail and review complete

+++ db/seeds.rb
@@ -1,19 +1,17 @@
 # frozen_string_literal: true
 
-require 'json'
 require 'open-uri'
-# This file should contain all the record creation needed to seed the database with its default values.
-# The data can then be loaded with the rails db:seed command (or created alongside the database with db:setup).
-#
-# Examples:
-#
-#   movies = Movie.create([{ name: 'Star Wars' }, { name: 'Lord of the Rings' }])
-#   Character.create(name: 'Luke', movie: movies.first)
+
+puts 'Destroy ingredients'
+Ingredient.destroy_all if Rails.env.development?
+
+puts 'Destroy Cocktails'
+Cocktail.destroy_all if Rails.env.development?
+
+puts 'Create ingredients'
 url = 'https://www.thecocktaildb.com/api/json/v1/1/list.php?i=list'
 ingredients = JSON.parse(open(url).read)
-
-puts 'Destroying previous data..'
-Ingredient.destroy_all
-ingredients['drinks'].each do |hash|
-  Ingredient.create(name: hash['strIngredient1'])
+ingredients['drinks'].each do |ingredient|
+  i = Ingredient.create(name: ingredient['strIngredient1'])
+  puts "create #{i.name}"
 end
